requirements.txt
evalutils==0.2.4
scikit-learn==0.20.2
scipy==1.2.1
--find-links https://download.pytorch.org/whl/torch_stable.html
torchvision==0.10.0+cu111
torchaudio==0.9.0
scikit-image==0.17.2
ensemble-boxes